+++ test/dateable_test.dart
@@ -4,15 +4,13 @@ import 'package:test/test.dart';
 void main() {
   group('Constructor tests', () {
     test('Date.fromDateTime(DateTime) constructor test', () {
-      expect(
-          Date.fromDateTime(DateTime(2002, 3, 11)), equals(Date(11, 3, 2002)));
+      expect(Date.fromDateTime(DateTime(2002, 3, 11)), equals(Date(11, 3, 2002)));
     });
     test('Date(int, int, int) date validation test', () {
       expect(Date(50, 40, 2000), equals(Date(20, 5, 2003)));
     });
     test('Date.parseIso8601(String) constructor test', () {
-      expect(Date.parseIso8601('2002-03-11T14:33:21.125'),
-          equals(Date(11, 3, 2002)));
+      expect(Date.parseIso8601('2002-03-11T14:33:21.125'), equals(Date(11, 3, 2002)));
     });
     test('Date.parse(String) constructor test', () {
       expect(Date.parse('11032002'), equals(Date(11, 3, 2002)));
@@ -30,12 +28,10 @@ void main() {
       expect(Date.today(), equals(DateTime.now().toDate()));
     });
     test('tomorrow constructor test', () {
-      expect(Date.tomorrow(),
-          equals(DateTime.now().add(const Duration(days: 1)).toDate()));
+      expect(Date.tomorrow(), equals(DateTime.now().add(const Duration(days: 1)).toDate()));
     });
     test('yesterday constructor test', () {
-      expect(Date.yesterday(),
-          equals(DateTime.now().subtract(const Duration(days: 1)).toDate()));
+      expect(Date.yesterday(), equals(DateTime.now().subtract(const Duration(days: 1)).toDate()));
     });
   });
   group('Comparision tests', () {
@@ -94,16 +90,13 @@ void main() {
       expect(Date(11, 3, 2002).isAfter(Date(11, 3, 2002)), equals(false));
     });
     test('copyWith test', () {
-      expect(Date(11, 3, 2002).copyWith(day: 21, month: 9),
-          equals(Date(21, 9, 2002)));
+      expect(Date(11, 3, 2002).copyWith(day: 21, month: 9), equals(Date(21, 9, 2002)));
     });
     test('isTheSameDate(Date) DateTime extension positive test', () {
-      expect(DateTime(2002, 3, 11, 14, 32, 56).isTheSameDate(Date(11, 3, 2002)),
-          equals(true));
+      expect(DateTime(2002, 3, 11, 14, 32, 56).isTheSameDate(Date(11, 3, 2002)), equals(true));
     });
     test('isTheSameDate(Date) DateTime extension negative test', () {
-      expect(DateTime(2004, 9, 21, 18, 26).isTheSameDate(Date(11, 3, 2002)),
-          equals(false));
+      expect(DateTime(2004, 9, 21, 18, 26).isTheSameDate(Date(11, 3, 2002)), equals(false));
     });
     test('isTheSameDate(Date) positive test', () {
       expect(Date(11, 3, 2002).isTheSameDate(Date(11, 3, 2002)), equals(true));
@@ -115,23 +108,16 @@ void main() {
       expect(DateTime.now().toDate().isToday(), equals(true));
     });
     test('isToday() negative test', () {
-      expect(DateTime.now().add(const Duration(days: 1)).toDate().isToday(),
-          equals(false));
+      expect(DateTime.now().add(const Duration(days: 1)).toDate().isToday(), equals(false));
     });
     test('isTomorrow() positive test', () {
-      expect(DateTime.now().add(const Duration(days: 1)).toDate().isTomorrow(),
-          equals(true));
+      expect(DateTime.now().add(const Duration(days: 1)).toDate().isTomorrow(), equals(true));
     });
     test('isTomorrow() negative test', () {
       expect(DateTime.now().toDate().isTomorrow(), equals(false));
     });
     test('isYesterday() positive test', () {
-      expect(
-          DateTime.now()
-              .subtract(const Duration(days: 1))
-              .toDate()
-              .isYesterday(),
-          equals(true));
+      expect(DateTime.now().subtract(const Duration(days: 1)).toDate().isYesterday(), equals(true));
     });
     test('isYesterday() negative test', () {
       expect(DateTime.now().toDate().isYesterday(), equals(false));
@@ -141,8 +127,11 @@ void main() {
     test('toDateTime() test', () {
       expect(Date(11, 3, 2002).toDateTime(), equals(DateTime(2002, 3, 11)));
     });
-    test('toIso8601() test', () {
+    test('toIso8601() with time test', () {
       expect(Date(11, 3, 2002).toIso8601(), equals('2002-03-11T00:00:00.000'));
+    });
+    test('toIso8601() without time test', () {
+      expect(Date(11, 3, 2002).toIso8601(includeTime: false), equals('2002-03-11'));
     });
     test('toString() test', () {
       expect(Date(2, 3, 104).toString(), equals('02030104'));
@@ -163,6 +152,25 @@ void main() {
     });
     test('subtractDays(int) test', () {
       expect(Date(7, 6, 2020).subtractDays(7), equals(Date(31, 5, 2020)));
+    });
+  });
+  group('weekday tests', () {
+    test('Weekday getter test 1', () {
+      expect(Date(11, 3, 2002).weekday, equals(1));
+    });
+    test('Weekday getter test 2', () {
+      expect(Date(21, 9, 2004).weekday, equals(2));
+    });
+  });
+  group('isValid tests', () {
+    test('isValid test 1', () {
+      expect(Date.isValid(day: 11, month: 3, year: 2002), equals(true));
+    });
+    test('isValid test 2', () {
+      expect(Date.isValid(day: 29, month: 2, year: 2021), equals(false));
+    });
+    test('isValid test 3', () {
+      expect(Date.isValid(day: 40, month: 50, year: -99999), equals(false));
     });
   });
 }